courses

+++ controllers/modulesController.js
@@ -1,30 +1,41 @@
-const ApiError = require("../exceptions/ApiError")
+const ApiError = require("../exceptions/ApiError");
+const lessonsService = require("../services/lessonsService");
 const moduleProgressService = require("../services/moduleProgressService")
 const modulesService = require("../services/modulesService")
+const coursesService = require("../services/coursesService")
 const roles = require("../utils/roles")
 
 class ModulesController {
     async createModule(req, res, next){
         try {
-            const moduleData = await modulesService.createModule(req.body)
-            res.json(moduleData)
+            const { role } = req.user;
+            const module = req.body;
+            if(role === roles.super){
+                await coursesService.getCourse(module.course)
+                const moduleData = await modulesService.createModule(module)
+                res.json(moduleData)
+            } else {
+                next(ApiError.Forbidden())
+            }
+
         } catch (e) {
             next(e)
         }
     }
-    async getModules(req, res, next){
-        try {
-            const data = await modulesService.getAllModules()
-            res.json(data)
-        } catch (e) {
-            next(e)
-        }
-    }
-    async getOneModule(req, res, next){
+    async getModule(req, res, next){
         try {
             const { id } = req.params;
-            const data = await modulesService.getOneModule(id)
-            res.json(data)
+            const { role, id: user } = req.user;
+            if(role === roles.user){
+                const Progress = await moduleProgressService.getProgress({ user, module: id })
+                const Module = await modulesService.getModuleLessonsProgress(id, user)
+                res.json({ ...Module, progress: Progress })
+            } else if(role === roles.super){
+                const Module = await modulesService.getModuleLessons(id)
+                res.json(Module)
+            } else {
+                next(ApiError.Forbidden())
+            }
         } catch (e) {
             next(e)
         }
@@ -32,8 +43,13 @@ class ModulesController {
     async updateModule(req, res, next){
         try {
             const { id } = req.params;
-            const data = await modulesService.updateModule(id, req.body)
-            res.json(data)
+            const { role } = req.user;
+            if(role === roles.super){
+                const data = await modulesService.updateModule(id, req.body)
+                res.json(data)
+            } else {
+                next(ApiError.Forbidden())
+            }
         } catch (e) {
             next(e)
         }
@@ -41,12 +57,20 @@ class ModulesController {
     async deleteModule(req, res, next){
         try {
             const { id } = req.params;
-            await modulesService.deleteModule(id)
-            res.json({message:"Запись о модуле удалена"})
+            const { role } = req.user;
+            if(role === roles.super){
+                await modulesService.deleteModule(id)
+                await lessonsService.deleteModuleLessons(id)
+                res.json({ message: "Модуль успешно удален" })
+            } else {
+                next(ApiError.Forbidden())
+            }
         } catch (e) {
             next(e)
         }
     }
+    
+
     async dropAllModules(req,res,next){
         try {
             const data = await modulesService.dropAllModules()
@@ -56,22 +80,5 @@ class ModulesController {
         }
     }
 
-    async getOneModuleLessons(req, res, next){
-        try {
-            const { id } = req.params;
-            const { role, id: user } = req.user;
-            if(role === roles.user){
-                const Module = await moduleProgressService.getModuleLessonseProgress(user, id);
-                res.json(Module)
-            } else if(role === roles.super){
-                const Module = await modulesService.getOneModuleData(id)
-                res.json(Module)
-            } else {
-                next(ApiError.Forbidden())
-            }
-        } catch (e) {
-            next(e)
-        }
-    }
 }
 module.exports = new ModulesController()
